#22. Fix compilation on MSVC.

+++ Executable/MemoryRecordingClient/Memory/Recording/Application/UI.hpp
@@ -10,7 +10,7 @@ public:
     void Render (class Client &_client) noexcept;
 
 private:
-    static uintptr_t ExtractGroupColor (const class RecordedAllocationGroup *_group,
+    static uint32_t ExtractGroupColor (const class RecordedAllocationGroup *_group,
                                         float _minBrightness,
                                         float _maxBrightness) noexcept;
